requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.0.5
sentence-transformers==2.2.2
faiss-cpu==1.12.0
numpy
requests
gunicorn
transformers
huggingface-hub
langchain
langdetect
